Remove unnecessary superclass
